codebase-ai-assistant/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
anthropic==0.38.0
supabase==2.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
GitPython==3.1.40
tree-sitter==0.21.3
tree-sitter-python==0.21.0
black==24.0.0
astroid==3.0.0
networkx==3.2
redis==5.0.1

